Tmux : change default-terminal to tmux-256color
- change `-g` to `-s`, since tmux version 2.1 `default-terminal`
is a server option, not a session option, as a side effect this
changes `default-terminal` to be a server rather than a session
option.

- change `default-terminal` value from `screen-256color` to
`tmux-256color`, because `screen-256color` entry in the
`terminfo` database doesn't have a `sitm` field. In other words,
the database reports that screen is unable to support italics,
and that is true, if `$TERM` set to `screen-256color`, when an
application need to make some text rendered italicized, it's
appear to be not possible, but it is possible, because it inside
tmux, not inside screen, and tmux does support the italics text
style, the solution is to set `$TERM` to `tmux-256color` so that
when an application queries the terminfo database, it finds the
field `sitm` with the right value `\E[3m`.