Neovim : from init.vim to init.lua
- move neovim config from `init.vim` to `init.lua` through
the following steps.

- remove `init.vim` file which was route neovim config to
`.vimrc` file ( vim config ).

- write neovim config completely in lua language because
neovim has an embedded lua runtime.

- why neovim use lua language.
lua language is used to create faster and more powerful
extensions for neovim editor, in the neovim charter, it lists
one of its goals as developing a first-class lua scripting
alternative to `viml`, one of the reasons for doing this is
that `viml` is a slow interpreted language with almost no
optimizations, much of the time spent in vim startup and in
actions from plugins that can block the main loop in the
editor is in parsing and executing vimscript.