Neovim : add nvimtree filesystem_watchers
- add `nvimtree` option `filesystem_watchers` to the
`config` table with type `table` and default value `{}`
the `filesystem_watchers` option will use file system
watcher (`libuv` `fs_event`) to watch the filesystem
for changes. using this will disable `BufEnter` and
`BufWritePost` events in `nvimtree` which were used to
update the whole tree. with this feature, the tree will be
updated only for the appropriate folder change, resulting
in better performance.