Neovim : add nvimtree actions.use_system_clipboard
- add `nvimtree` option `actions.use_system_clipboard` to
the `config` table with type `boolean` and default value
`true` the `actions.use_system_clipboard` option a boolean
value that toggle the use of system clipboard when `copy`
or `paste` function are invoked. When enabled, copied text
will be stored in registers '+' (system), otherwise, it will
be stored in '1' and '"'.